Infosys_Project/project1/requirements.txt
streamlit
praw
pandas
requests
python-dotenv
python-docx
PyPDF2
